+++ nvim/lua/user/plugins.lua
@@ -149,6 +149,32 @@ require("lazy").setup({
         enabled = true,
     },
     {
+        -- Autocompletion
+        "hrsh7th/nvim-cmp",
+        dependencies = {
+            "hrsh7th/cmp-nvim-lsp",
+            "hrsh7th/cmp-buffer",
+            "hrsh7th/cmp-path",
+        },
+        config = function()
+            local cmp = require("cmp")
+            cmp.setup({
+                sources = {
+                    { name = "nvim_lsp" },
+                    { name = "buffer" },
+                    { name = "path" },
+                },
+                mapping = cmp.mapping.preset.insert({
+                    ["<C-Space>"] = cmp.mapping.complete(),
+                    ["<CR>"] = cmp.mapping.confirm({ select = false }),
+                    ["<Tab>"] = cmp.mapping.select_next_item(),
+                    ["<S-Tab>"] = cmp.mapping.select_prev_item(),
+                }),
+            })
+        end,
+        enabled = true,
+    },
+    {
         -- nice pop selection menu for the UI
         "nvim-telescope/telescope-ui-select.nvim",
         dependencies = {